// client/src/types/Order.ts
import { CartItem, ShippingAddress } from "./Cart"
import { User } from "./User"

export type Order = {
    _id: string
    orderItems: CartItem[]
    shippingAddress: ShippingAddress
    paymentMethod: string
    user: User
    createdAt: string
    isPaid: boolean
    paidAt: string
    isDelivered: boolean
    deliveredAt: string
    itemsPrice: number
    shippingPrice: number
    taxPrice: number
    totalPrice: number
}

export type createRazorpay = {
amount:string
amount_due:number
amount_paid:number
attempts:number
created_at:number
currency:string
entity:string
id:string
    
}